+++ src/http/service/object-service.js
@@ -1,5 +1,6 @@
 import { ObjectSchema } from '../../schemas/object/object-schema.js';
 import { ApiError } from '../errors/api-error.js';
+import mongoose from 'mongoose';
 
 class ObjectService {
   async create({ name, lat, lon, country, region, city, street, building, room, description }) {
@@ -16,6 +17,12 @@ class ObjectService {
   async read(id) {
     if (!id) {
       throw ApiError.BadRequest({ message: 'Need id for return object' });
+    }
+
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid Object id" });
     }
 
     const findSuchObject = await ObjectSchema.findById(id).populate('numSDSs');
@@ -36,6 +43,12 @@ class ObjectService {
       throw ApiError.BadRequest({ message: 'Has not required fields: name, lat, lon, country or region' });
     }
 
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid Object id" });
+    }
+
     const updatedObject = await ObjectSchema.findByIdAndUpdate(
       id,
       { name, lat, lon, country, region, city, street, building, room, description },
@@ -52,6 +65,12 @@ class ObjectService {
   async delete(id) {
     if (!id) {
       throw ApiError.BadRequest({ message: 'Need id for delete object' });
+    }
+
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid Object id" });
     }
 
     const deletedObject = await ObjectSchema.findByIdAndDelete(id);
