+++ src/http/service/sds-service.js
@@ -1,5 +1,6 @@
 import { SdsSchema } from '../../schemas/sds/sds-schema.js';
 import { ApiError } from '../errors/api-error.js';
+import mongoose from 'mongoose';
 
 class SDSService {
   async create({ name, model, phone, comment, objectId }) {
@@ -8,14 +9,20 @@ class SDSService {
     }
 
     const SDS = new SdsSchema({ name, model, phone, comment, objectId });
-
     await SDS.save();
+
     return SDS;
   }
 
   async read(id) {
     if (!id) {
       throw ApiError.BadRequest({ message: 'Need id for return SDS' });
+    }
+
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid SDSs id" });
     }
 
     const findSuchSDS = await SdsSchema.findById(id).populate('numDevices');
@@ -28,12 +35,18 @@ class SDSService {
   }
 
   async update(id, { name, model, phone, comment, objectId }) {
+    if (!id) {
+      throw ApiError.BadRequest({ message: 'Has not SDSs id for update' });
+    }
+
     if (!name || !model || !phone || !comment || !objectId) {
       throw ApiError.BadRequest({ message: 'Has not required fields: name, model, phone, comment or objectId' });
     }
 
-    if (!id) {
-      throw ApiError.BadRequest({ message: 'Has not SDSs id for update' });
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid SDSs id" });
     }
 
     const updatedSDS = await SdsSchema.findByIdAndUpdate(id, { name, model, phone, comment, objectId }, { new: true });
@@ -48,6 +61,12 @@ class SDSService {
   async delete(id) {
     if (!id) {
       throw ApiError.BadRequest({ message: 'Need id for delete SDS' });
+    }
+
+    const isValidId = mongoose.Types.ObjectId.isValid(id);
+
+    if (!isValidId) {
+      throw ApiError.BadRequest({ message: "It's not a valid SDSs id" });
     }
 
     const deletedSDS = await SdsSchema.findByIdAndDelete(id);
